Kursa ogrenci ekleme islemlemleri tamamlandi

+++ javaprojects/src/com/company/bolum_13_collections_project/udemykursplayer/Ders.java
@@ -2,17 +2,17 @@ package com.company.bolum_13_collections_project.udemykursplayer;
 
 public class Ders {
 
-    private int dersNo;
+    private double dersNo;
     private String dersBaslik;
     private double dakika;
 
-    public Ders(int dersNo, String dersBaslik, int dakika) {
+    public Ders(int dersNo, String dersBaslik, double dakika) {
         this.dersNo = dersNo;
         this.dersBaslik = dersBaslik;
         this.dakika = dakika;
     }
 
-    public int getDersNo() {
+    public double getDersNo() {
         return dersNo;
     }
